+++ example/lib/src/view.dart
@@ -5,6 +5,9 @@
 // The Material design platform for Flutter
 export 'package:flutter/material.dart';
 
+/// Access to the localizationsDelegates
+export 'package:flutter_localizations/flutter_localizations.dart';
+
 export 'package:example/src/app/view.dart';
 
 export 'package:example/src/home/view.dart';